requirements.txt
APScheduler==3.8.1
art==5.1
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.4
django-allauth==0.44.0
django-apscheduler==0.6.0
django-cors-headers==3.10.0
django-model-utils==4.1.1
djangorestframework==3.12.4
ExifRead==2.3.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
Unipath==1.1
urllib3==1.26.3
wfastcgi==3.0.0
wrapt==1.12.1
xmltodict==0.12.0